Drawing textures
Using SOIL, a texture was loaded, and glTexture was used to render the texture to the screen. In the vertices array, UV coordinates were added to map the texture correctly to the verteices. The shaders were adapted to make use of Sampler2D and adding images ot the rendered primitives